#195 Refactor: rename file and variable
