+++ javascript/big-practice/src/scripts/components/formAddNote.js
@@ -1,7 +1,7 @@
 import iconPin from '../../assets/icons/icon-pin.svg';
 import iconColorBoard from '../../assets/icons/icon-color-board.svg';
 
-const formTemplate = () => {
+const formAddNote = () => {
   const formControl = document.createElement('div');
   formControl.setAttribute('class', 'note-wrapper');
 
@@ -36,4 +36,4 @@ const formTemplate = () => {
   return formControl;
 };
 
-export default formTemplate;
+export default formAddNote;
