+++ javascript/big-practice/src/scripts/components/inputSearch.js
