+++ javascript/big-practice/src/scripts/components/listNotes.js
@@ -1,7 +1,7 @@
 import iconTrash from '../../assets/icons/icon-in-trash.svg';
 import iconLight from '../../assets/icons/icon-light.svg';
 
-const noteTemplate = (note) => {
+const listNotesWrapper = (note) => {
   const wrapper = document.createElement('div');
   wrapper.setAttribute('class', `${note.type === 'trashNotes' ? 'trash-wrapper' : 'list-cover'}`);
   wrapper.innerHTML = `
@@ -18,4 +18,4 @@ const noteTemplate = (note) => {
   return wrapper;
 };
 
-export default noteTemplate;
+export default listNotesWrapper;
