+++ javascript/big-practice/src/scripts/components/logo.js
